requirements.txt
discord.py==2.4.0
numpy==2.2.1
pandas==2.2.3
